testing email_course

+++ app/routes/api/v1/files.tsx
@@ -29,25 +29,18 @@ export const action = async ({ request }: Route.ActionArgs) => {
   }
   if (intent === "delete_file") {
     const url = new URL(request.url);
-    const serviceURL = new URL(`https://video-converter-hono.fly.dev`);
-    // const storageKey = url.searchParams.get("storageKey")!;
-    serviceURL.pathname = "/delete_all";
-    serviceURL.searchParams.set(
-      "webhook",
-      "https://easybits.cloud/api/v1/conversion_webhook"
-    );
-    // serviceURL.searchParams.set("storageKey", storageKey);
-    // const response = await fetch(serviceURL.toString(), {
-    //   method: "delete",
-    //   headers: {
-    //     Authorization: "Bearer PerroTOken", // @todo from service dashboard
-    //   },
-    // });
-    // const text = await response.text();
-    // console.info("::CONVERTER_RESPONSE::", text, response.status);
-    // if (response.status === 404) {
+    // const serviceURL = new URL(`https://video-converter-hono.fly.dev`);
+    // // const storageKey = url.searchParams.get("storageKey")!;
+    // serviceURL.pathname = "/delete_all";
+    // serviceURL.searchParams.set(
+    //   "webhook",
+    //   "https://easybits.cloud/api/v1/conversion_webhook"
+    // );
     const storageKey = (url.searchParams.get("storageKey") ||
       formData.get("storageKey")) as string;
+
+    if (!storageKey) throw new Response("File not found", { status: 404 });
+
     await deleteObject(storageKey); //Revisit
     try {
       await db.file.delete({
